Add quickjs submodule for Micro-Application-Runtime (MAR) support
Introduce the quickjs submodule under 3rd/ directory to enable
integration with the Micro-Application-Runtime project. This submodule
facilitates MAR's usage of QuickJS, a lightweight JavaScript engine. The
addition of this submodule streamlines the process of including QuickJS
in MAR projects and enables seamless collaboration with the QuickJS
repository. This commit sets up the submodule and adds the necessary
configuration to the .gitmodules file.